refactor cli to use cobra framework
