+++ go.mod
@@ -13,8 +13,10 @@ require (
 	github.com/jinzhu/copier v0.3.2
 	github.com/onsi/ginkgo v1.14.1
 	github.com/onsi/gomega v1.10.2
+	github.com/spf13/cobra v1.1.1
 	github.com/stretchr/testify v1.7.0
 	google.golang.org/grpc v1.36.0
+	k8s.io/api v0.20.2
 	k8s.io/apimachinery v0.20.2
 	k8s.io/client-go v0.20.2
 	sigs.k8s.io/controller-runtime v0.8.3
